Improve TimeController, use Slot, fix history/realtime display times, fix problem switching play/pause by using onlick event on div etc.
